Fixed scaling function bug
By default, the data are scaled such that values are between 0 and 1 during estimation (for stability). After running the model, the solution values (parameters, gradient, and hessian) are appropriately re-scaled. There was a bug in the code where if one alternative was an outside good (i.e. all attributes = 0) it would throw an error. This is now resolved.